requirements.txt
jupyterlab==4.0.8
pyro-ppl==1.8.6
scikit-learn==1.3.2
seaborn==0.13.0
torch==2.1.0
torchvision==0.16.0
